% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scda.test.R
\docType{package}
\name{scda.test-package}
\alias{scda.test}
\alias{scda.test-package}
\title{scda.test: Testing \code{tern} and \code{scda} Data}
\description{
This package provides a testing suite for templates based on
\code{scda} data and \code{tern} analysis methods.
}
\details{
To learn mode about the package, check the documentation of \link[tern:tern]{tern}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/scda.test/}
  \item Report bugs at \url{https://github.com/insightsengineering/scda.test/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com}

Authors:
\itemize{
  \item Joe Zhu \email{joe.zhu@roche.com}
  \item Daniel Sabanés Bové \email{daniel.sabanes_bove@rconis.com}
  \item Jana Stoilova \email{stoilova@dnli.com}
  \item Heng Wang \email{Heng.Wang21@gilead.com}
  \item Francois Collin
  \item Adrian Waddell \email{aadrian.waddell@gmail.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Chendi Liao
  \item Jennifer Li
  \item Jan Szczypiński \email{jan.szczypinski@external.roche.com} (\href{https://orcid.org/0000-0002-5682-5840}{ORCID})
}

Other contributors:
\itemize{
  \item David Munoz Tord \email{david.munoztord@mailbox.org} [contributor]
  \item Johnson & Johnson Innovative Medicine [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
