% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scda.test.R
\docType{package}
\name{scda.test-package}
\alias{scda.test}
\alias{scda.test-package}
\title{scda.test: Testing \code{tern} and \code{scda} Data}
\description{
This package provides a testing suite for templates based on
\code{scda} data and \code{tern} analysis methods.
}
\details{
To learn mode about the package, check the documentation of \link{tern}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/scda.test/}
  \item Report bugs at \url{https://github.com/insightsengineering/scda.test/issues}
}

}
\author{
\strong{Maintainer}: insightsengineering \email{insightsengineering@example.com}

}
\keyword{internal}
